Enforce the account window, drop idle binds, filter by status, sliding throttle
- A session with more unanswered submits than its account window gets
  ESME_RMSGQFUL, recorded in the inbox as rejected. Visible with latency
  rules or fast clients, which is when real operators return it
- Binds silent for CELLOPHANE_IDLE_TIMEOUT (default 2m; 0 disables) are
  closed via Netty's IdleStateHandler; enquire_link keeps them alive
- status= filter (comma-separated, case-insensitive) on the messages
  list and a status dropdown in the UI; live updates respect the filter
- throttle counts a sliding one-second window instead of wall-clock
  seconds, so a burst across a second boundary cannot pass twice the tps
- README screenshot regenerated with the logo and the filter; docs and
  changelog updated. 135 tests

+++ server/src/main/java/io/cellophane/server/operator/Metrics.java
@@ -26,7 +26,7 @@ public final class Metrics {
         submitted.incrementAndGet();
         (wasAccepted ? accepted : rejected).incrementAndGet();
         long second = clock.instant().getEpochSecond();
-        int slot = (int) Math.floorMod(second, WINDOW_SECONDS);
+        int slot = Math.floorMod(second, WINDOW_SECONDS);
         if (perSecondStamp.get(slot) != second) {
             perSecondStamp.set(slot, second);
             perSecond.set(slot, 0);
